Some fiddling with log formatting
One of the problems with info logging is that the user frequently
doesn't want that annoying "INFO:" prefix getting in the way of the data
they're seeing.  This change tries shoving the metadata out of the way
by right justifying it using the display size.

Add an example to test out the style interactively.