Improve lazy log formatting and key=value handling
Rework logmsg macro so that it wraps the log formatting (including
creation of key value pairs) in a closure to enable generic filtering
before log formatting, and to get a lot of the complex formatting code
off the main code path.

To go with this, introduce the idea of key=value log control hints,
allowing hints for filtering which are eagerly evaluated and passed to a
newly reworked `shouldlog()` function in a standard way.

In addition, protect user code from exceptions thrown during log
formatting so that enabling debugging can't cause an application to
crash.

Also add a whole bunch of documentation for the logmsg macro.

TODO:
* Extra tests for new functionality