Workaround for unbox poisioning due to closure capture
Add a find_var_uses() utility function to find variable uses.  Rebind
these variables with `let` to work around poisioning of inference by
closure capture https://github.com/JuliaLang/julia/issues/15276.

This is ugly as it essentially implements some subset of the lowering
rules (and is therefore probably still buggy as julia has a rather rich
syntax).

It is, however, very effective and entirely necessary in julia-0.6 if we
want to use a closure for the purpose of deferring log formatting.