Refactor custom log level machinery
LogLevel is now a custom type (rather than an enum) to give a little extra
control, and allow a range of nonstandard log levels to be represented
with the LogLevel type.  This also clarifies the custom log level
interface a bit, in that custom levels only need to support conversion
to LogLevel in order to work with the fast log limiting machinery.