Rework log formatting, rendering log strings as markdown
Render all log strings as markdown by default.  This means logging will
be able to support multi line messages with meaningful formatting for
really readable user interaction.

This could cause formatting havoc if data is to be logged, but in that
case the user defined key value pairs are a better option in any case.

Other changes:

* Fix text justification issues by having a single coloured prefix
  character for each log line, and keeping metadata on the right.

* Decide on `exception` key for conveying exceptions, rather than using
  a tuple.