Value passthrough and improved result printing for test_logs
Pass the value of the test expression along from test_logs and
test_deprecated, so that the value can be tested using an additional
call to @ test.

@ test_logs now records a LogTestFailure with the current test set, which
means we have some control over how the offendig logs are printed.
Unfortunately this also means this part of MicroLogging is dependent on
implementation details of Base.Test (this problem can only really be
fixed by refactoring Test itself).